split evaluate into its own suite and remove pin from models
