+++ .github/workflows/master.yml
@@ -248,11 +248,33 @@ jobs:
       - name: Install dependencies
         run: |
           source ./dev/install-common-deps.sh
-          pip install pyspark "numpy<1.24.0"
+          pip install pyspark
       - name: Run tests
         run: |
           export MLFLOW_HOME=$(pwd)
           pytest tests/models
+
+  evaluate:
+    runs-on: ubuntu-latest
+    timeout-minutes: 90
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          submodules: recursive
+          repository: ${{ github.event.inputs.repository }}
+          ref: ${{ github.event.inputs.ref }}
+            - uses: ./.github/actions/setup-python
+            - uses: ./.github/actions/setup-pyenv
+            - uses: ./.github/actions/setup-java
+            - name: Install dependencies
+              run: |
+                source ./dev/install-common-deps.sh
+                pip install pyspark "numpy<1.24.0"
+            - name: Run tests
+              run: |
+                export MLFLOW_HOME=$(pwd)
+                pytest tests/evaluate
+          
 
   pyfunc:
     runs-on: ubuntu-latest
