+++ tests/evaluate/test_validation.py
@@ -15,10 +15,6 @@ from unittest import mock
 import pytest
 
 # pylint: disable=unused-import
-from tests.models.test_evaluation import (
-    multiclass_logistic_regressor_model_uri,
-    iris_dataset,
-)
 
 message_separator = "\n"
